# apps/api/openapi.gen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Char AI API",
    "description": "AI services API for speech-to-text transcription, LLM chat completions, and subscription management",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/pyannote/v1/diarize": {
      "post": {
        "tags": [
          "pyannote"
        ],
        "summary": "Diarize audio",
        "operationId": "diarize",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiarizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreated"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "402": {
            "description": "Subscription is required"
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "x-api-key-permissions": [
          "jobs:write"
        ]
      }
    },
    "/pyannote/v1/identify": {
      "post": {
        "tags": [
          "pyannote"
        ],
        "summary": "Identify speaker with diarization",
        "operationId": "identify",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreated"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "402": {
            "description": "Subscription is required"
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "x-api-key-permissions": [
          "jobs:write"
        ]
      }
    },
...
    "/pyannote/v1/media/input": {
      "post": {
        "tags": [
          "pyannote"
        ],
        "summary": "Get upload URL",
        "description": "To use the provided temporary storage is a two step process.\n    You start by declaring a media:// url that you can reference in any other API calls. The response will provide a url where you can put your media. This allows you to use the media:// url as a short-cut for a temporary storage location.\n    You'll be returned a pre-signed url you can use to PUT and upload your media file. The temporary storage should allow you to read and write to the media:// locations for a period of at least 24 hours before it is removed.",
        "operationId": "getMediaUploadURL",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetMediaUploadUrl"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "402": {
            "description": "Subscription is required"
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "x-api-key-permissions": [
          "media:write"
        ]
      }
    },
...
    "/pyannote/v1/voiceprint": {
      "post": {
        "tags": [
          "pyannote"
        ],
        "summary": "Extract voiceprint",
        "operationId": "voiceprint",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceprintRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreated"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "402": {
            "description": "Subscription is required"
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ],
        "x-api-key-permissions": [
          "jobs:write"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiarizeRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "confidence": {
            "type": "boolean",
            "description": "Include confidence values in the output. Output is considerably larger when this option is enabled. Output includes a list of confidence scores with a resolution.",
            "default": false,
            "example": true
          },
          "exclusive": {
            "type": "boolean",
            "description": "Includes exclusive diarization values in the output in `exclusiveDiarization` key (equivalent to diarization but without overlapping speech).",
            "default": false,
            "example": true
          },
          "maxSpeakers": {
            "type": "number",
            "description": "Maximum number of speakers (must be >= minSpeakers if both are set)",
            "example": 4,
            "minimum": 1
          },
          "minSpeakers": {
            "type": "number",
            "description": "Minimum number of speakers (must be <= maxSpeakers if both are set)",
            "example": 1,
            "minimum": 1
          },
          "model": {
            "type": "string",
            "default": "precision-2",
            "enum": [
              "precision-2",
              "community-1"
            ],
            "example": "precision-2"
          },
          "numSpeakers": {
            "type": "number",
            "description": "Number of speakers. Only use if the number of speakers is known in advance. Number of speakers is detected automatically if not provided. Setting this value results in better overall diarization performance. In rare cases where we cannot honor this request (e.g. short files and large number of speakers), a warning will be added to the output. Equivalent to sending minSpeakers==maxSpeakers",
            "example": 2,
            "minimum": 1
          },
          "transcription": {
            "type": "boolean",
            "description": "Enable speaker attributed transcription. Only available for the `precision-2` diarization model.",
            "default": false
          },
          "transcriptionConfig": {
            "$ref": "#/components/schemas/TranscriptionConfiguration"
          },
          "turnLevelConfidence": {
            "type": "boolean",
            "description": "Includes turn-level confidence values in the output.",
            "default": false,
            "example": true
          },
          "url": {
            "type": "string",
            "description": "URL of the audio file to be processed",
            "example": "https://example.com/audio.wav"
          },
          "webhook": {
            "type": "string",
            "description": "Webhook URL to receive results when job is completed (optional)",
            "example": "https://example.com/webhook"
          },
          "webhookStatusOnly": {
            "type": "boolean",
            "description": "When true, webhook payload only includes jobId and status (excludes output). Useful for large payloads.",
            "default": false,
            "example": true
          }
        }
      },
...
      "GetMediaUploadUrl": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string",
            "description": "The url should be in the form media://object-key where the object-key can be any alpha-numeric string. The object-key is unique to your account API token so there is no risk of collision with other users.",
            "maxLength": 255,
            "pattern": "^media:\\/\\/[a-zA-Z0-9\\-_\\.\\/]+$"
          }
        }
      },
...
      "IdentifyRequest": {
        "type": "object",
        "required": [
          "voiceprints"
        ],
        "properties": {
          "confidence": {
            "type": "boolean",
            "description": "Include confidence values in the output. Output is considerably larger when this option is enabled. Output includes a list of confidence scores with a resolution.",
            "default": false,
            "example": true
          },
          "exclusive": {
            "type": "boolean",
            "description": "Includes exclusive diarization values in the output in `exclusiveDiarization` key (equivalent to diarization but without overlapping speech).",
            "default": false,
            "example": true
          },
          "matching": {
            "$ref": "#/components/schemas/MatchingOptions"
          },
          "maxSpeakers": {
            "type": "number",
            "description": "Maximum number of speakers (must be >= minSpeakers if both are set)",
            "example": 4,
            "minimum": 1
          },
          "minSpeakers": {
            "type": "number",
            "description": "Minimum number of speakers (must be <= maxSpeakers if both are set)",
            "example": 1,
            "minimum": 1
          },
          "model": {
            "type": "string",
            "default": "precision-2",
            "enum": [
              "precision-2"
            ],
            "example": "precision-2"
          },
          "numSpeakers": {
            "type": "number",
            "description": "Number of speakers. Only use if the number of speakers is known in advance. Number of speakers is detected automatically if not provided. Setting this value results in better overall diarization performance. In rare cases where we cannot honor this request (e.g. short files and large number of speakers), a warning will be added to the output. Equivalent to sending minSpeakers==maxSpeakers",
            "example": 2,
            "minimum": 1
          },
          "turnLevelConfidence": {
            "type": "boolean",
            "description": "Includes turn-level confidence values in the output.",
            "default": false,
            "example": true
          },
          "url": {
            "type": "string",
            "description": "URL of the audio file to be processed",
            "example": "https://example.com/audio.wav"
          },
          "voiceprints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Voiceprint"
            },
            "description": "List of voiceprints to identify against",
            "maxItems": 50,
            "minItems": 1
          },
          "webhook": {
            "type": "string",
            "description": "Webhook URL to receive results when job is completed (optional)",
            "example": "https://example.com/webhook"
          },
          "webhookStatusOnly": {
            "type": "boolean",
            "description": "When true, webhook payload only includes jobId and status (excludes output). Useful for large payloads.",
            "default": false,
            "example": true
          }
        }
      },
...
      "JobCreated": {
        "type": "object",
        "required": [
          "jobId",
          "status"
        ],
        "properties": {
          "jobId": {
            "type": "string",
            "description": "ID of the job",
            "example": "3c8a89a5-dcc6-4edb-a75d-ffd64739674d"
          },
          "status": {
            "type": "string",
            "description": "Status of the job",
            "enum": [
              "pending",
              "created",
              "succeeded",
              "canceled",
              "failed",
              "running"
            ],
            "example": "created"
          },
          "warning": {
            "type": "string",
            "description": "Warning message if any"
          }
        }
      },
...
      "MatchingOptions": {
        "type": "object",
        "properties": {
          "exclusive": {
            "type": "boolean",
            "description": "Prevent multiple speakers from being matched to the same voiceprint. Default to true",
            "default": true
          },
          "threshold": {
            "type": "number",
            "format": "float",
            "description": "Prevent matching if confidence score is below this threshold. Value is between 0 and 100. Default is 0, meaning all voiceprints are matched",
            "default": 0,
            "maximum": 100,
            "minimum": 0
          }
        }
      },
      "MediaResponse": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
...
      "TranscriptionConfiguration": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string",
            "description": "Transcription model to use. Supported languages depend on the chosen model:\n  - `parakeet-tdt-0.6b-v3`: Bulgarian, Croatian, Czech, Danish, Dutch, English, Estonian, Finnish, French, German, Greek, Hungarian, Italian, Latvian, Lithuanian, Maltese, Polish, Portuguese, Romanian, Slovak, Slovenian, Spanish, Swedish, Russian or Ukrainian.\n  - `faster-whisper-large-v3-turbo`: Afrikaans, Albanian, Amharic, Arabic, Armenian, Assamese, Azerbaijani, Bashkir, Basque, Belarusian, Bengali, Bosnian, Breton, Bulgarian, Cantonese, Catalan, Chinese, Croatian, Czech, Danish, Dutch, English, Estonian, Faroese, Finnish, French, Galician, Georgian, German, Greek, Gujarati, HaitianCreole, Hausa, Hawaiian, Hebrew, Hindi, Hungarian, Icelandic, Indonesian, Italian, Japanese, Javanese, Kannada, Kazakh, Khmer, Korean, Lao, Latin, Latvian, Lingala, Lithuanian, Luxembourgish, Macedonian, Malagasy, Malay, Malayalam, Maltese, Maori, Marathi, Mongolian, Myanmar, Nepali, Norwegian, Nynorsk, Occitan, Pashto, Persian, Polish, Portuguese, Punjabi, Romanian, Russian, Sanskrit, Serbian, Shona, Sindhi, Sinhala, Slovak, Slovenian, Somali, Spanish, Sundanese, Swahili, Swedish, Tagalog, Tajik, Tamil, Tatar, Telugu, Thai, Tibetan, Turkish, Turkmen, Ukrainian, Urdu, Uzbek, Vietnamese, Welsh, Yiddish or Yoruba.",
            "default": "parakeet-tdt-0.6b-v3",
            "enum": [
              "parakeet-tdt-0.6b-v3",
              "faster-whisper-large-v3-turbo"
            ],
            "example": "faster-whisper-large-v3-turbo"
          }
        }
      },
...
      "Voiceprint": {
        "type": "object",
        "required": [
          "label",
          "voiceprint"
        ],
        "properties": {
          "label": {
            "type": "string",
            "description": "Label for the speaker. Labels can't start with \"SPEAKER_\"",
            "example": "John Doe",
            "maxLength": 100,
            "pattern": "^(?!speaker_).*"
          },
          "voiceprint": {
            "type": "string",
            "format": "base64",
            "description": "Voiceprint of a speaker",
            "example": "U29tZUJhc2U2NERhdGE",
            "maxLength": 20000
          }
        }
      },
...
      "VoiceprintRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "model": {
            "type": "string",
            "default": "precision-2",
            "enum": [
              "precision-2"
            ],
            "example": "precision-2"
          },
          "url": {
            "type": "string",
            "description": "URL of the voiceprint audio file",
            "example": "https://example.com/voice.wav"
          },
          "webhook": {
            "type": "string",
            "description": "Webhook URL to receive voiceprint results (optional)",
            "example": "https://example.com/webhook"
          },
          "webhookStatusOnly": {
            "type": "boolean",
            "description": "When true, webhook payload only includes jobId and status (excludes output). Useful for large payloads.",
            "default": false,
            "example": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token"
      }
    }
  },
...
}